move logger into separate package
